requirements.txt
streamlit==1.28.1
tensorflow-cpu==2.15.0
numpy==1.24.3
scikit-learn==1.3.0
Pillow==10.0.1
pandas==2.0.3
protobuf==3.20.3
h5py==3.9.0
scipy==1.11.3